requirements.txt
requests
python-dotenv
httpx
numpy
web3
